+++ packages/adminui-vue/components/controls/UCheckbox.vue
@@ -12,7 +12,10 @@
       class="u-checkbox__label"
       :class="{ 'u-checkbox__label--left': labelPosition === 'left' }"
       :for="_uid"
-    >{{ $ut(label) }}</label>
+    >{{ $ut(label) }}
+      <!-- @slot Default slot is inside label, just after input -->
+      <slot />
+    </label>
   </span>
 </template>