+++ packages/adminui-vue/CHANGELOG.md
@@ -6,8 +6,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - `UCheckbox` - added default slot - allows insert some content inside a checkbox label
 
 ### Changed
+ - `UCron` - rewrite:
+   - added non-standard 7-part cron expression syntax, where last seven field `@n` mean - fires on every `n` occurrence
+   - added "Estimated executions" - first 5 dates what match cron expression.
+     Uses server-side `ubq_scheduler.estimateCronSchedule` method (if available)
+   - added two-way binding support
+   - added hints for cron expression fields of `Expresion`
+   - correct implementation of `step` modifier - allow construct field like `1-10/3`, `3,4,5,6/2` etc.
+   - *BREAKING* `UCron` do no longer emits `change` event with expression + it's explanation.
+     `$App.verbaliseCronExpression(value)` should be used to get expression explanation. 
 
 ### Deprecated